requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
catalog-config==1.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
requests==2.27.1
sendgrid==6.9.7
six==1.16.0
soupsieve==2.3.2.post1
starkbank-ecdsa==2.0.3
urllib3==1.26.9
